requirements.txt
requests==2.31.0
python-telegram-bot==20.7
backoff==2.2.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0